+++ ex_app/lib/all_tools/audio2text.py
@@ -3,6 +3,7 @@
 from langchain_core.tools import tool
 from nc_py_api import Nextcloud
 
+from ex_app.lib.all_tools.lib.context import get_nextcloud
 from ex_app.lib.all_tools.lib.files import get_file_id_from_file_url
 from ex_app.lib.all_tools.lib.task_processing import run_task
 from ex_app.lib.all_tools.lib.decorator import safe_tool
@@ -18,6 +19,8 @@ async def get_tools(nc: Nextcloud):
 		:param file_url: The file URL to the media file in nextcloud (The user can input this using the smart picker for example)
 		:return: the transcription result
 		"""
+		nonlocal nc
+		nc = get_nextcloud(nc)
 		task_input = {
 			'input': get_file_id_from_file_url(file_url),
 		}
