+++ ex_app/lib/all_tools/files.py
@@ -1,10 +1,12 @@
 # SPDX-FileCopyrightText: 2025 Nextcloud GmbH and Nextcloud contributors
 # SPDX-License-Identifier: AGPL-3.0-or-later
+from fastmcp.server.dependencies import get_context
 from langchain_core.tools import tool
 from nc_py_api import Nextcloud
 
 from typing import Optional
 
+from ex_app.lib.all_tools.lib.context import get_nextcloud
 from ex_app.lib.all_tools.lib.decorator import safe_tool, dangerous_tool
 
 
@@ -17,7 +19,9 @@ async def get_tools(nc: Nextcloud):
 		Get the content of a file
 		:param file_path: the path of the file
 		:return: 
-		""" 
+		"""
+		nonlocal nc
+		nc = get_nextcloud(nc)
 
 		user_id = nc.ocs('GET', '/ocs/v2.php/cloud/user')["id"]
 		
@@ -34,7 +38,9 @@ async def get_tools(nc: Nextcloud):
 		Get the folder tree of the user
 		:param depth: the depth of the returned folder tree
 		:return: 
-		""" 
+		"""
+		nonlocal nc
+		nc = get_nextcloud(nc)
 
 		return nc.ocs('GET', '/ocs/v2.php/apps/files/api/v1/folder-tree', json={'depth': depth}, response_type='json')
 
@@ -45,7 +51,9 @@ async def get_tools(nc: Nextcloud):
 		Creates a public sharing link for a file or folder
 		:param path: the path of the file or folder
 		:return: 
-		""" 
+		"""
+		nonlocal nc
+		nc = get_nextcloud(nc)
 
 		response = nc.ocs('POST', '/ocs/v2.php/apps/files_sharing/api/v1/shares', json={
 					'path': path,
